Fix comm count microbenchmarks
Previously, these tests always ran the default backend (labelled as
C-backend) and if llvm was enabled it would run llvm variants as well.
With the switch to llvm as our default and changes made in 17800 we
stopped running the default config for these microbenchmarks and only
ran llvm-wideopt variants.

Fix that here by renaming the configs to avoid mention of the backend
and be more explicit about cache vs. no-cache and wideopt vs. no-wideopt
to avoid this problem with future changes to our default config.

Resolves Cray/chapel-private 2409

Signed-off-by: Elliot Ronaghan <ronawho@gmail.com>