Fix use-after-free bug in regex chpl__serialize
in _serialize, qio_regex_get_pattern returnas a copy of the pattern
string, which we then store a pointer to in the serialize helper. The
function returns and we free that buffer because it was an owned
buffer. _deserialize then tries to do a copy of the now-freed buffer
and we have a use-after-free.

The fix included switches to use a borrowed string gotten from the new
qio_regex_borrow_pattern. That buffer is valid as long as the regex is
alive, which is long enough for the matching _deserialize function to
work.

qio_regex_borrow_pattern could have been introduced in addition to
qio_regex_get_pattern, but the number of call sites is low and
create{String,Bytes}WithNewBuffer seems natural

Signed-off-by: Andrew Consroe <andrew.consroe@hpe.com>