Update compiler/next handling of Location
1) Add ID to ErrorMessage

ErrorMessage previously only stored a Location. We'd like to store an
ID so that if a query result is re-used after an AST has moved in a
file (but not changed), the ErrorMessage stored wants a way to get
updated with the current information.

The idea here is to use an explicit update function, passing in the
Context as that is necessary for getting the current location.

---

2) Add tests for compiler/next incremental locations

  a) Previously, errors could be reported with old line numbers. Add a
  test to verify an error message uses the new line number.

  b) Previously, locateAst would also return old line numbers. Add a
  test to verify that locating both AST and Comment nodes return the
  correct location

---

3) Add CommentID and map locations by ID and CommentID

This adds a CommentID used similar to ID, but we want to separate the
notion that comments don't participate in the result of queries.

CommentID is set on a Comment

BuilderResult now maps ID -> Location for all AST nodes except Comment
BuilderResult now maps CommentID -> Location for all Comment

The "map" of CommentID -> Location is really a vector, because the
numbering of Comment's is dense.

A CommentID does not store the module or file it came from, so the
caller should only look up the location for a Comment with the
BuilderResult it came from. This is a reasonable expectation for use
in chpldoc for example.

Signed-off-by: Andrew Consroe <andrew.consroe@hpe.com>