Fix invalid reuse of partial result take 2
The previous approach of clearing the result after deciding to recompute
the query doesn't work because it leads to the comparison after the query
to always using a new result (which mean queries that depend on it will
be recomputed).

Instead, in this commit, switch to having a separate partialResult
field in the query entry, and clear this in a new
QUERY_END_CURRENT_RESULT() that also does some swapping around
that was happening in a few places.

(Work was by mppf and I picked half a commit)

---
Signed-off-by: Michael Ferguson <mppf@users.noreply.github.com>
Signed-off-by: Andrew Consroe <andrew.consroe@hpe.com>