Fix problem in get_compiler_from_command
PR #18745 assumed that all calls to it would be passing in just one
command (since the arguments are frequently now stored in a list and so
call sites can just get the first list element). However that's not the
case in get_compiler_from_cc_cxx where this function is called with the
value of CC / CXX environment variables.

---
Signed-off-by: Michael Ferguson <mppf@users.noreply.github.com>