Add additional protection against indexing into heterogeneous tuples
The test shown here resulted in an internal rather than user-facing
error, I _believe_ due to the fact that it was following a speculative
resolution path in trying to determine the type of an expression and
squashing user errors along the way which would have flagged this
expression as being illegal.  Here, I took the simple approach of
replicating the normal module-level user-facing error message in place
of the assert that was being tripped prior to this change.

---
Signed-off-by: Brad Chamberlain <bradcray@users.noreply.github.com>