Fix a problem with dynamic dispatch in the parallel pass
When creating heap allocations during the parallel pass, a dynamically
dispatched method call should be handled the same as other function calls.
Add it to a branch that is handling resolved calls instead of handling it
alongside random primitive calls.

Removed the special future for this case that was only running with --no-local
configurations, and allow the other identical copy to run in both --local and
--no-local configurations.

Signed-off-by: David Iten <daviditen@users.noreply.github.com>