Deprecate iostyle and another function that returned it
Deprecates `stringStyleWithLength`, replaces the name iostyle in the runtime
with `iostyleInternal` and defines a new, private extern record in IO by that
same name.  Deprecates iostyle in IO and makes it a type alias that points to
the new extern record type.  Replaces a link to an iostyle field with just the
name of it (it's in a deprecated function's documentation, so I'm not going to
do more with it).  Updates a bunch of extern declarations to use the new extern
record, but there's still more references to iostyle that need to be updated

----
Signed-off-by: Lydia Duncan <lydia-duncan@users.noreply.github.com>