Resolve a test that started breaking on case-insensitive platforms
Lydia pointed out in issue #18821 that when we renamed Regexp.chpl to
Regex.chpl, tests in this directory that had been doing 'use Regexp;'
were updated to 'use Regex;' which was disambiguated to the test named
'regex.chpl' in this directory.  Here, I'm just renaming the test to
regexp.chpl to avoid the ambiguity.

---
Signed-off-by: Brad Chamberlain <bradcray@users.noreply.github.com>