Issue error if mkdtemp fails
Previously, if mkdtemp failed to create the directory requested, it
returns NULL which got propagated up, resulting in a segfault when it
tried to be used.

This change checks the return value and issues a fatal error instead.

Also dup'ing the return value of astr because we shouldn't be
modifying it.

Thanks to mppf for the suggestion on using a second astr

Signed-off-by: Andrew Consroe <andrew.consroe@hpe.com>