Remove Makefile.freebsd
It was added in #4629 but that change is causing
it to find a system-installed RE2 rather than the bundled
one. The problem is that /usr/local/include and /usr/local/lib
are searched before the paths to elements of our runtime.

FreeBSD compilers don't automatically search /usr/local/
even though that is by default where packages are installed.

This commit removes that change from #4629 which means that
some libcurl applications will no longer compile on FreeBSD.
However, the better solution for that (and other similar dependencies)
is to use mason / pkg-config to figure out what paths are needed.
Additionally we need to make sure that any bundled paths occur
before any system-wide paths in the compile line.

---
Signed-off-by: Michael Ferguson <mppf@users.noreply.github.com>