Simplify blocking-AM management.
The part of the common Active Message support that dealt with whether or
not we block waiting for the AM to finish was unnecessarily complicated,
especially for AM-mediated atomic ops where we might want to delay that
blocking until our memory consistency model requires the atomic op to be
done.  Here, refactor it so that the callers simply tell the common code
whether or not to block immediately, and handle the delayed-blocking
case entirely in the AM-mediated AMO code.  As part of this, bring the
delayed-blocking set-up inline into the AM-mediated AMO code -- it's
small and only used in that one place, and things are more clear if
everything is right there.

Signed-off-by: Greg Titus <gbtitus@users.noreply.github.com>