Follow up fix for ordered outputs
Additional fix for #18870

This test was still dependent on the hash order because there was an
array of type `[1..3] [enumDom] real` and the inner array wasn't being
output in sorted order.

I've only implemented this change in the one file because it is the
only one that uses nested arrays like this.

Signed-off-by: Andrew Consroe <andrew.consroe@hpe.com>