Add util/devel/nextLinter
This is a linter to help properly implementing some conventions in
compiler/next

It uses the JSON dump of the C++ AST

Currently it flags unused fields in operator==, mark, and hash.

It has an allow-list for unused fields of primitve type (int, enum,
etc) in mark.

It is rather slow (5-10s), consumes plenty of memory (~1G), and requires clang
12+ (I get weird segfaults with 11).

Signed-off-by: Andrew Consroe <andrew.consroe@hpe.com>