Wrap UniqueString temporaries with astr
This fixes an ASAN stack-use-after-scope that is triggered by passing
the const char* that points into a UniqueString temporary (presumably
when it is in small string mode).

This happens because a call like `->name()` returns a UniqueString by
value (creating a temporary) and is only alive until the end of the
expression (after the `.c_str()` in this case).

Another fix would be to assign into a local variable like

UniqueString tmp = as->rename()->name();
const char* rename = tmp.c_str();
...

This fix is not ideal b/c any constructor that we pass one of these
into is going to call `astr` itself, so it's a bit of wasted effort. I
view this as a temporary fix to quiet testing before break.

Signed-off-by: Andrew Consroe <andrew.consroe@hpe.com>