Switch {array,domain,distribution}.targetLocales to parenless
This PR switches the `targetLocales` method on Chapel arrays, domains,
and distributions to be paren-less. It implements the decision
made by the domain module review on 6/14/21 (internal issue 2118)
and extends it to arrays and distributions for consistency.

This adjusts the compiler to emit a warning when such targetLocales
is invoked with parentheses. This was the easiest way to implement
such a warning. This adjustment should be removed after the deprecation
period.

This adjusts modules and tests to implement this change and
adds a test for the warning in test/deprecated.

Future work: add `targetLocales` to the default distribution.

Signed-off-by: Vassily Litvinov <vasslitvinov@users.noreply.github.com>