Add a constraints file to arkouda testing itself for now
This is a hotfix for not having previously pinned setuptools versions.  Release
arkouda testing will not otherwise respond to changes to version numbers.  We
should remove this hotfix in the future.

Adds a constraints file to the arkouda testing directory, and uses it when
installing python testing requirements to explicitly limit which version of
setuptools is used.

----
Signed-off-by: Lydia Duncan <lydia-duncan@users.noreply.github.com>