Make first-class functions preserve return intents
Prior to this commit, the compiler's implementation of first-class
functions as class methods failed to preserve the function's return
intent on the method, as pointed out by @thomasrolinger on issue
making 'ref' and 'const ref' intents work as intended.

I've added a future here indicating that we don't properly complain
about functions that return 'const ref' being stored in 'ref'
variables, which is orthogonal to this issue but something I
stumbled across in testing.

---
Signed-off-by: Brad Chamberlain <bradcray@users.noreply.github.com>