Fix support for chains of implicit module uses and require of .chpl files
BenMcD pointed out that we were not doing a good job of parsing files
in require statements that were in implicitly used modules.  This was
due to the lack of a loop in the logic that starts parsing dependent
files that weren't explicitly named on the command-file.  The fix was
fairly simple: add a loop that continues parsing things while it still
finds new things that need to be parsed.  This PR does so and also adds
an extended version of Ben's motivating test that takes the chain one
step further to make sure that it can spin for multiple times (I got
it wrong the first time around).

---
Signed-off-by: Brad Chamberlain <bradcray@users.noreply.github.com>