Clean up `pthread_once()` use in the runtime.
We use `pthread_once()` in a number of places in the runtime.  In some
cases we've been guarding it with a boolean so as to avoid calling it
when we could tell we had already done so.  But in at least one of those
cases we mis-coded the guard, setting it and then assigning the set-once
value, in that order, rather than assigning the set-once value, then
doing a fence, then setting the guard.  In addition, examining at least
the glibc `nptl` version of pthread_once() that we use on Linux systems
shows that it effectively does the same thing, using fast-path control
flow to avoid any locking when the called-once function is known to have
completed already.  It seems likely that other implementations would do
similarly.  Therefore here, remove all our guards on `pthread_once()`
calls.

While here, add a missing include of `chpl-atomics.h` in the ofi comm
layer, which I spotted while working on this.

Signed-off-by: Greg Titus <gbtitus@users.noreply.github.com>