Add missing braces
std::array needed another pair of braces when -Wmissing-braces was
on. When assigning with = you can drop the outermost one, but not here

Adds -Wmissing-braces to the compiler CXX flags so
we (I!) don't hit this in the future. Only added to gnu and clang as
those will be used in development (and these are warnings enabled with
CHPL_DEVELOPER)

Fixes a build error reported by the gasnet team. Thanks!

Note that it seems clang doesn't report this warning, but I've still
added it to Makefile.clang

Signed-off-by: Andrew Consroe <andrew.consroe@hpe.com>