chpldoc/next - now with queries
Squash of progress on chpldoc/next overhaul to use queries. That was
an easy step.

Adds an initial implementation of actually converting the AST into
the appropriate RST format. This uses the visitor and traverse
pattern.

Added a hack environment variable to sub_test.py so that
CHPL_CHPLDOC_NEXT can be specified to use it instead of today's
chpldoc. This was because PATH get's messed with by fixpath.py

Note this "requires" C++17 only as a convenience step in prototyping
to get the fileystem functionality.

Today we only pass 1 test case. Many are whitespace only which is a
good sign, just ran out of time to address them.

Signed-off-by: Andrew Consroe <andrew.consroe@hpe.com>