Move some method definitions into headers
These small method definitions are being moved into their header files
so that they are eligible for inlining, which most of these almost
certainly will.

This was motivated by seeing these methods commonly show up in the top
20 of profiling reports:

  * Symbol::hasFlag
  * CallExpr::isPrimitive
  * Expr::inTree
  * AList::get
  * Symbol::isRef

I used a bit of loose judgement to move some related functions if they
felt sufficiently small and/or similar to one of the above. I also
chose to keep the method definitions outside the class as it makes a
better diff.

I had to move the class definition for Expr into expr-class-def.h so
that the alist.h header can implement the ::get method. Expr contains
an AList so it also needs to know the full layout. Without major
changes to header structure, this was my solution.

I verified most of these actually do get inlined and saw a 3-5%
reduction in compiling hello, cg-sparse-timecomp, and arkouda.

Signed-off-by: Andrew Consroe <andrew.consroe@hpe.com>