Compute rst documentation as a query
This extends the previous work of adding queries to next/chpldoc by
converting the one-shot visitor approach with a qury rstDoc to compute
the docs for a given node. These results form a tree with the goal
being that the query framework will re-use subtrees that have not
changed. This avoids a small blow-up in eagerly concatenating a tree
of strings but the main motivation was to kick the tires more on
queries.

This now passes a whopping 15 chpldoc tests! 138 Failures still :)

Also fixes a small bug in mark-functions.h

Signed-off-by: Andrew Consroe <andrew.consroe@hpe.com>