Don't always add -O3 to flags when compiling with gasnet
Follow-up to PR #18880.

The pkg-config .pc file we get gasnet flags from includes -O3 in the
flags, which caused a failure for the test

  expressions/bradc/uminusVsTimesPrec

(which is sensitive to optimization level and skipped with --fast).

Previously the Makefile based approach avoided the -O3 by getting
only particular kinds of flags. We could do that too with the pkg-config
parsing as well but for now, just filter out the flags we don't want
after we parse them.

---
Signed-off-by: Michael Ferguson <mppf@users.noreply.github.com>