Fix failing tests
This adds the ability to use a test.tasks-{none,qthreads,fifo}.good
file in the test system and documents it along with a few scenarios
missing from the docs.

Most failing tests were fixed with the _QUIET=yes env var so they
aren't chatty and removed the need for a prediff in a few cases.

Signed-off-by: Andrew Consroe <andrew.consroe@hpe.com>