Better error diagnostics for unsupported domain operations
This PR implements better compile-time error messages in cases
when a common domain operation is invoked on a domain that
does not support it. For example, when `stride` is invoked
on an associative domain, the compiler now reports
`this domain type does not support stride`.

The covered operations are:

```
low
high
stride
alignment
first
last
alignedLow
alignedHigh
parSafe
indexOrder
orderToIndex
bulkAdd
makeIndexBuffer
```

Operations that have already provided this kind of diagnostics
and so are not affected by this PR include:
`stridable` `expand` `exterior` `interior` `translate`

The implementation is based on adding "last resort" overloads
that generate compilerError. This way domain maps are free
to add implementations of these operations.

SOME RELEVANT TRIVIA

"Last resort" methods take precedence during resolution over
forwarded methods. Therefore this PR adds some method definitions
that do the forwarding "manually", esp. in ArrayViewRankChange.chpl
and ArrayViewReindex.chpl, to avoid the error-generating overloads.

Using the `override` keyword causes a compilation error when the names
of the formal(s) in the overriding method and the overridden method differ.
Therefore some methods that became overrides as a result of this PR
do not have the `override` annotation.

WHILE THERE

Remove some "The following procedure is effectively equivalent to"
comments as they are no longer helpful.

Improve _makeIndexTuple() to add specificity to the error message,
when it is generated upon operations like `by` and `align`.

Improve runtime error reporting in `orderToIndex`.

NEXT STEPS

It will be a usability improvement if the error message reports
the type mismatch between the available function and its callsite
in the case an operation is available on a given domain however it is
invoked with incorrect arguments. Currently the existing compiler machinery
that does such reporting in other cases "misfires" in the context
of this PR: the error message highlights a candidate function that is not
applicable because its where clause evaluates to false, rather than
the candidate that would apply except for the actual/formal type mismatch.
Example operations include `bulkAdd` and `expand`.

Enable testing for sparse subdomains of associative domains, see
`dfltAssociativeSparseSub`, once we implement such subdomains.

Enable the testing in notsupportedRectOps-ok.chpl for `DistributedBase`
and `DistributedSubdomain` once Block domains start supporting
some standard rectangular-domain operations like `first`.

Signed-off-by: Vassily Litvinov <vasslitvinov@users.noreply.github.com>