Initial commit of printchplbuilds.
printchplbuilds is a tool that displays information about the available
runtime builds. It is helpful when chpl complains that a runtime for
your current configuration is not available. It has options for
sorting the builds, changing how the builds are displayed, emitting
shell commands to match a particular build, and checking that a build
exists that matches the current configuration.

Signed-off-by: John H. Hartman <jhh67@users.noreply.github.com>