Add initial compiler/next query timing and tracing
For performance profiling of compiler/next queries, this adds two
capabilities:

1) `Context::setQueryTimingFlag(true)` records the time taken in
  a) getMap
  b) getResult
  c) the query function body (timer starts after getResult)
2) `Context::{begin,end}QueryTimingTrace("/tmp/out.txt")` enables logging
the queries executed with their full callgraph structure (tree
suffices)

The query timing trace (quite a mouthful to say) has a postprocessor
analyze-query-trace that reconstructs the tree, computes time in self,
summarizes, and can plot the results in DOT

The timings themselves are stored inside the QueryMapBase because
these stats are small and already accessible in the query body.

This is excercised in `testQueryTimingAndTrace`, but it is like
`testInteractive` in that it is only really run interactively for now.

Signed-off-by: Andrew Consroe <andrew.consroe@hpe.com>