Remove 257threads test
This test continued to fail after changes from #18871
and would need the same fixes as in #19046

That fix made it possible to handle this correctly, supporting
.tasks-* files that differentiate the warnings (since they differ
whether you use fifo or qthreads). However, this test is made
redundant by ManyThreads, which additionally tests with comm=none, so
this coverage is not contributing. It's sibiling test, the -quiet
version, does add coverage because it is testing the --quiet execopt

Signed-off-by: Andrew Consroe <andrew.consroe@hpe.com>