Update the compiler to support LLVM-13
Update the compiler with changes required to work with LLVM-13. Maintain
backward compatibility with versions 12 and 11.

The widest change was changing calls to the CreateLoad method to take the
value's type as its the first argument.

There were also a handful of flags that changed availability when passing
options between Clang and LLVM, functions that were deprecated and other
small change.

Signed-off-by: David Iten <daviditen@users.noreply.github.com>