Report an error for a non-rectangular sparse domains
Currently sparse subdomains are implemented only for rectangular domains,
including sparse subdomains of (simple and sparse) subdomains with
a rectangular domain as the final transitive parent.

This PR makes it a compile-time error to create a sparse subdomain type
that does not fit the above category. This way users will not encounter
unhelpful error messages when invoking some functionality (such as bulkAdd()
currently) on a non-rectangular sparse domain.

Signed-off-by: Vassily Litvinov <vasslitvinov@users.noreply.github.com>