Report an error for first/last on empty sparse domains
Currently spec leaves the behavior of `.first` / `.last`
on empty domains unspecified. Such queries on non-sparse
rectangular domains return low/high, whereas on sparse domains
they produce an "array index out of bounds" error, noting that
"index was 1 but array bounds are 1..0".

Switch the latter behavior to an error message that indicates directly
that first/last are invoked on an empty domain.

Signed-off-by: Vassily Litvinov <vasslitvinov@users.noreply.github.com>