Report errors for improper args of 'sparse subdomain'
Report an error upon `sparse subdomain(x)` where x is a type or not a domain.

This was motivated by my writing this innocent-looking code, which used to
produce an error message that is unhelpful and reveals implementation details:

    var DR = [1..3];
    var SS: sparse subdomain(DR);

To enable this, move some AST building logic from the parser into the modules.

Signed-off-by: Vassily Litvinov <vasslitvinov@users.noreply.github.com>