Deprecate+better errors for | & ^ op= on rectangular domains
This PR implements an action item from the domain module review,
where we decided to disallow `|` `&` `^` on rectangular domains.
For this release these are deprecated. We will convert them
to compilation errors in a future release, unless we receive
opposing requests from users.
Given that `|=` `&=` `^=` on rectangular domains already generate
compile-time errors, this PR makes these error messages clear.

It adds tests of the deprecated warnings and the improved error messages.
A couple of .good/.bad files are adjusted also to account for a change
in program AST and the number of available candidates.

While there: move += and -= on (domain,idx) from ChapelBase into ChapelDomain
as that is where they belong.
Archaeology: they were added initially in cd9b794b5f
and had been in ChapelBase since then.

Signed-off-by: Vassily Litvinov <vasslitvinov@users.noreply.github.com>