Update references to SSCA2 in the GRAPHFILES
Uses new location, and performs some resorting - now that the test no longer
lives in release/examples, it shouldn't be part of the `Release examples` suite.
We also don't want it to be part of the `Performance tracking` suite since we
don't feel as confident in it.  Elliot suggested moving it to the `Proxy apps`
suite (and renaming the suite so it is no longer DOE specific).

----
Signed-off-by: Lydia Duncan <lydia-duncan@users.noreply.github.com>