Change line number searched for spectral-norm verification
The output of these spectral norm variants now includes a deprecation message.
As a result, the line number where the validation output is generated is now
pushed below this deprecation message, which was causing our perfkeys validation
to fail.

The IO subteam expects to have a new strategy for adjusting the output in place
of using the iostyle type and style arguments for various functions.  Once that
new strategy is in place, we will need to revert this change (as the test will
stop generating a deprecation message)

----
Signed-off-by: Lydia Duncan <lydia-duncan@users.noreply.github.com>