In comm=ugni, sanity-check transfer extents against their apparent MRs.
We've been assuming in comm=ugni that if the starting address of a
transfer was found to be within some memory registration, there were
sufficient constraints elsewhere in the Chapel software stack to
guarantee the whole extent of the transfer would lie within that MR.  We
still believe that should be true for Chapel itself, but we've recently
encountered a situation in which not-strictly-Chapel code made it false.
Specifically, turning on chunk merging and splitting in the jemalloc
memory allocation package caused us to end up with arrays that spanned
uGNI memory registration boundaries, leading to error returns from uGNI
that are hard to diagnose because the specific error code we get back
isn't pretty nonspecific.  We've fixed that problem, but in addition to
that, here adjust the comm layer to sanity-check the whole extent of
each transfer against the MR containing its starting address.  This will
give us a specific error for this problem if something else causes it in
the future.

Signed-off-by: Greg Titus <gbtitus@users.noreply.github.com>