Reduce amount of testing of day21a
This test is highly recursive and copies tiny arrays on each function
call, so is pretty brutal with memtracking on.  Skip the test when
that's the case as a result.  In addition, the two configurations are
testing virtually the same thing and take ~15-20 seconds on a Mac or
linux64, so let's just run one of them (where the "non-practice" run
tends to be faster in my experience).

---
Signed-off-by: Brad Chamberlain <bradcray@users.noreply.github.com>