Use hybrid MR mode with the cxi provider
As described by @gbtitus:

The libfabric cxi provider for Cassini-based networks now supports the
so-called "hybrid" local memory registration mode. For functions that
accept a local MR descriptor, this allows the client to pass the
non-NULL MR descriptor (pointer) if the local address is already
registered, or NULL to indicate that the local memory is not registered
and the provider should use its internal MR cache. (Absent this hybrid
mode the provider always uses its internal MR cache, which is a source
of contention for multi-threaded clients such as Chapel and in certain
cases SHMEM.)

Signed-off-by: John H. Hartman <jhh67@users.noreply.github.com>
Signed-off-by: slnguyen <26969019+slnguyen@users.noreply.github.com>