Write "internal error" messages to `stderr`, even in debug-comm mode.
We had been writing "internal error" messages into the debug file when the
comm layer was built in comm-debug mode.  But that support was broken,
resulting in no output, when the user didn't set any debug flags.  Here, just
go back to printing such messages to stderr so that they appear in the same
place with or without debug.  (But still flush output and do an `abort()` in
debug mode, so we get a core file if that's enabled.)

Signed-off-by: Greg Titus <gbtitus@users.noreply.github.com>