When checking object extents against MRs, allow some slack in the length.
When we create memory registration table entries they always have starting
addresses that are aligned, either to base page boundaries for certain of the
static memory segments we collect from `/proc/self/maps`, or to hugepage
boundaries for the fixed heap or any dynamic memory segments.  However, we
don't round their lengths upward; we just record the length the caller gave
us.  So although the actual registration is certainly an integral number of
pages in length, our recorded length is only the usually-smaller amount the
user asked for.  As it turns out, comm cache prefetching (or maybe even plain
old fetching that is not "pre", I'm not sure) can access up to a cache line
length, less one byte, past the end of a registration when pulling data from a
remote node.  It was therefore tripping the new extent check when such a cache
fetch extended beyond the user's requested registration length, even though
the "excess" bytes were indeed within the registration itself and the prefetch
was perfectly safe.  Here, deal with this by rounding the MR length up to the
next cache line length when doing the extent check.

The other way to do this would be to round up the length at the time we do the
registration.  However, I think for accuracy we'd want to round up to the page
boundary rather than a cache line, so that our recorded length actually was
the registration length.  Unfortunately we don't know the page size in the
common code that does the registration and MR management, and getting that
right would be a larger change than I want to take on in this PR.  And of
course another alternative would be to round up the MR length by a page size
rather than a cache line at the time of the extent check.  But leaving the
constraint just loose enough to allow current behavior to work seems like the
right thing to do, because it will force us to look at this again if something
changes and we start (pre)fetching more than just a cache line.

Signed-off-by: Greg Titus <gbtitus@users.noreply.github.com>