Fix simple mistake pointed out by Damian
In discussing math stuff recently, Damian pointed out that Chapel's
cproj() function returned (and assumed C was returning) reals/floats
rather than complexes.  This makes the simple update to address that.

(Damian also suggested that cproj() is so simple that we ought to
implement it in Chapel (from a simplicity / optimizability of code
perspective), but I did not feel I had the time to take that up late
on a Friday.  Someone should open an issue about this or take it on
themselves if they feel strongly about it.

---
Signed-off-by: Brad Chamberlain <bradcray@users.noreply.github.com>