Report a clear error when using min/max(type t)
When invoking `min` or `max` on a type that does not define it,
report an error stating so instead of a generic "unresolved call" error.
This is reasonable given that these are predefined operations in Chapel.

The error overloads added here are marked with the "last resort" pragma,
so users are free to add their own overloads that will take priority.

Resolves #6312

Signed-off-by: Vassily Litvinov <vasslitvinov@users.noreply.github.com>