First draft of ChplConfig module
This draft comments out all the variables in ChapelEnv.chpl because
it's the easiest way to debug issues that show up in the internal
modules, which is where most of these uses take place.  If that
code is uncommented, the user gets a warning about the need to
use/import ChplConfig in the future.

---
Signed-off-by: Brad Chamberlain <bradcray@users.noreply.github.com>